preparing the release
