+++ Heartland-iOS-SDK/Classes/Terminals/UPA/Builders/HpsUpaAuthBuilder.h
@@ -18,7 +18,6 @@
 @property (nonatomic, strong) NSString *transactionId;
 
 // HSA/FSA Values
-@property (nonatomic, readwrite) BOOL cardIsHSAFSA;
 @property (nonatomic, strong) NSDecimalNumber *prescriptionAmount;
 @property (nonatomic, strong) NSDecimalNumber *clinicAmount;
 @property (nonatomic, strong) NSDecimalNumber *dentalAmount;
