+++ Heartland-iOS-SDK/Classes/Terminals/UPA/Builders/HpsUpaVerifyBuilder.h
@@ -19,9 +19,6 @@
 @property (nonatomic, readwrite) HpsStoredCardInitiator storedCardInitiator;
 @property (nonatomic, strong) NSString *cardBrandTransactionId;
 
-// HSA/FSA Values
-@property (nonatomic, readwrite) BOOL cardIsHSAFSA;
-
 - (void) execute:(void(^)(HpsUpaResponse*, NSError*))responseBlock;
 - (id)initWithDevice: (HpsUpaDevice*)upaDevice;
