+++ Heartland-iOS-SDK/Classes/Terminals/UPA/Models/HpsUpaTransaction.h
@@ -14,7 +14,6 @@
 @property (nonatomic,retain)NSString* referenceNumber;
 @property (nonatomic,retain)NSString* tranNo;
 @property (nonatomic,retain)NSString* processCPC;
-@property (nonatomic,retain)NSString* cardIsHSAFSA;
 @property (nonatomic,retain)NSString* prescriptionAmount;
 @property (nonatomic,retain)NSString* clinicAmount;
 @property (nonatomic,retain)NSString* dentalAmount;
