--- Heartland-iOS-SDK/Frameworks/GlobalMobileSDK.xcframework/ios-arm64/GlobalMobileSDK.framework/PrivateHeaders/BBDeviceCAPK.h
@@ -1,29 +1,0 @@
-//
-//  BBDeviceCAPK.h
-//
-//  Created by Alex Wong on 2017-08-18.
-//  Copyright © 2020 BBPOS Limited. All rights reserved.
-//  RESTRICTED DOCUMENT
-//
-
-#import <Foundation/Foundation.h>
-
-@interface BBDeviceCAPK : NSObject{
-    NSString *location;
-    NSString *index;
-    NSString *size;
-    NSString *rid;
-    NSString *modulus;
-    NSString *exponent;
-    NSString *checksum;
-}
-
-@property (nonatomic, retain) NSString *location;
-@property (nonatomic, retain) NSString *index;
-@property (nonatomic, retain) NSString *size;
-@property (nonatomic, retain) NSString *rid;
-@property (nonatomic, retain) NSString *modulus;
-@property (nonatomic, retain) NSString *exponent;
-@property (nonatomic, retain) NSString *checksum;
-
-@end
